Change dir structure and reorder components
