+++ src/components/common/Button/Button.css
@@ -1,9 +1,3 @@
-nav {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
 button {
     font-family: var(--font-text-primary);
     font-size: 1.7em;
