+++ src/components/common/Button/Button.js
@@ -1,0 +1,12 @@
+import "./Button.css";
+
+export function Button({
+    text,
+    onClickHandler,
+}) {
+    return(
+        <button onClick={ onClickHandler }>
+            { text }
+        </button>
+    )
+}
