+++ src/components/common/Card/Card.css
@@ -1,0 +1,10 @@
+.card {
+    background-color: var(--dark-secondary);
+    border-radius: 10px;
+    display: flex;
+}
+
+.card-body {
+    padding: 15px;
+    width: 100%;
+}
