+++ src/components/common/Card/index.js
@@ -1,0 +1,1 @@
+export { Card } from './Card';
