+++ src/components/common/Modal/Modal.css
@@ -1,0 +1,17 @@
+.modal-base {
+    height: 100vh;
+    width: 100vw;
+    display: flex;
+    position: absolute;
+    top: 0;
+    left: 0;
+    background-color: rgba(0, 0, 0, 0.9);
+}
+
+.modal-content {
+    width: 90%;
+    margin: 80px auto auto auto;
+    padding: 10px 15px;
+    background-color: var(--dark-secondary);
+    border-radius: 10px;
+}
