+++ src/components/common/Modal/index.js
@@ -1,0 +1,1 @@
+export { Modal } from './Modal';
