+++ src/components/common/Navbar/Navbar.css
@@ -1,0 +1,5 @@
+nav {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
