+++ src/components/common/Navbar/index.js
@@ -1,0 +1,1 @@
+export { Navbar } from './Navbar';
