+++ src/components/common/WorldMap/index.js
@@ -1,0 +1,1 @@
+export { WorldMap } from './WorldMap';
