+++ src/components/specific/LocationSearchModal/index.js
@@ -1,0 +1,1 @@
+export { LocationSearchModal } from './LocationSearchModal';
