+++ src/components/specific/LocationTime/index.js
@@ -1,0 +1,1 @@
+export { LocationTime } from './LocationTime';
