+++ src/components/specific/TimeZoneCard/index.js
@@ -1,0 +1,1 @@
+export { TimeZoneCard } from './TimeZoneCard';
